This commit was manufactured by cvs2svn to create tag 'WX_2_3_4'.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/tags/WX_2_3_4@17994 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/mac/carbon/printmac.cpp
@@ -276,7 +276,7 @@ void wxMacPrintPreview::DetermineScaling(void)
 	
 	m_previewPrintout->SetPPIScreen( 72 , 72 ) ;
 	m_previewPrintout->SetPPIPrinter( 72 , 72 ) ;
-	m_previewPrintout->SetPageSizeMM( 8 * 25.6 , 11 * 25.6 ) ;
+	m_previewPrintout->SetPageSizeMM( (int) (8.0 * 25.6), (int) (11.0 * 25.6) );
 	m_previewPrintout->SetPageSizePixels( 8 * 72 , 11 * 72 ) ;
     m_pageWidth = 8 * 72 ;
     m_pageHeight = 11 * 72 ;
